Data Structures for Parser
Add some data structures need in implementing the deterministic parser.

1. Trie
--
A non-generic trie implementation for efficient storage of utterances
and their mapped outputs.
The implementation is influenced by the `trie` and `radix-trie` crates:
- [trie](https://github.com/contain-rs/trie)
- [radix trie](https://github.com/michaelsproul/rust_radix_trie)

2. Symbol Table
--
A non-generic table for storing string symbols, this basically
a wrapper around Rust's stdlib's HashMap with additional hashing
for generating symbol keys.

There is also a corresponding Foreign Function Interface to expose
the data structures apis to the C FFI.